added prolog program for overapproximation of simple type analysis
To determine the bounds for the symmetry reduction, we need to know how many
objects from a given object set can ever occur together in an instantiation of a
schematic operator that might become relaxed reachable. For tight bounds, the
program computes an overapproximation of whether a operator parameter x could
be from the given object set in the relevant situations.